Readd vipe method to Diskstorage because it will be useful for timing and testing code. Move compressed versions to their own file.
