+++ .gitignore
@@ -132,3 +132,6 @@ dmypy.json
 .idea
 /storage
 /.vscode
+
+/examples/storage
+/examples/storage2
